refactor: schedule 도메인 디렉토리 위치 이동
