+++ apps/user-service/src/main/java/site/icebang/domain/schedule/mapper/ScheduleMapper.java
@@ -1,10 +1,10 @@
-package site.icebang.schedule.mapper;
+package site.icebang.domain.schedule.mapper;
 
 import java.util.List;
 
 import org.apache.ibatis.annotations.Mapper;
 
-import site.icebang.schedule.model.Schedule;
+import site.icebang.domain.schedule.model.Schedule;
 
 @Mapper
 public interface ScheduleMapper {
