+++ apps/user-service/src/main/resources/mybatis/mapper/ScheduleMapper.xml
@@ -1,16 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="site.icebang.schedule.mapper.ScheduleMapper">
+<mapper namespace="site.icebang.domain.schedule.mapper.ScheduleMapper">
 
-    <select id="findAllByIsActive" resultType="site.icebang.schedule.model.Schedule"> SELECT
+    <select id="findAllByIsActive" resultType="site.icebang.domain.schedule.model.Schedule"> SELECT
                                                                                                          id AS scheduleId,
                                                                                                          workflow_id AS workflowId,
                                                                                                          cron_expression AS cronExpression,
                                                                                                          is_active AS isActive
-                                                                                      FROM
+                                                                                             FROM
                                                                                                          schedule
-                                                                                      WHERE
+                                                                                             WHERE
                                                                                                          is_active = #{isActive}
     </select>
